TouchID Implementation [iOS Sim Only] (#509)
* Fixed broken tests

Tests were failing because of an indentation problem in
IOSGesturesTest.java and because a static import in
XCUIAutomationTest.java was coming after the rest of the imports instead
of before

* Added 'touch_id' endpoint

(See https://support.apple.com/en-ca/HT201371 for description of the Touch ID feature)
This endpoint simulates the TouchID feature.

* removed obselete gson-2.6.2 dependency. (#507)

* Fix documentation

* Fix documentation

* Don't commit 'touchID()' 

This will be added post release

* Added touchId test

This test only executes the driver.touchId method, it doesn't do any assertions. Just verifies that it doesn't throw any exceptions.

* Added assertion to keep codacy happy

* Fixed broken tests

Tests were failing because of an indentation problem in
IOSGesturesTest.java and because a static import in
XCUIAutomationTest.java was coming after the rest of the imports instead
of before

* Added 'touch_id' endpoint

(See https://support.apple.com/en-ca/HT201371 for description of the Touch ID feature)
This endpoint simulates the TouchID feature.

* Fix documentation

* Fix documentation

* Don't commit 'touchID()' 

This will be added post release

* Added touchId test

This test only executes the driver.touchId method, it doesn't do any assertions. Just verifies that it doesn't throw any exceptions.

* Added assertion to keep codacy happy

* Moved touchID into new file

-Moved touchID out of IOSDeviceActionShorcuts and into PerformsTouchID
-Made IOSDriver implement PerformsTouchID
-Fixed typo that I noticed in CommandExecutionHelper
-Fixed linting errors

* Fixed conflicts

* Update IOSMobileCommandHelper.java

* Throws exception instead of bogus assertion