Update gradle build script (#803)
* Upgrade to Gradle 4.4.1

* Upgrade to Shadow plugin 2.0.2 to avoid using deprecated API

Build with Gradle 4.4.1 produces warning "A constructor for `org.gradle.api.internal.java.JavaLibrary` is used by Shadow plugin v1.2.x, and has been preserved for compatibility. This has been deprecated and is scheduled to be removed in Gradle 5.0. If you're using the Shadow plugin, try upgrading to v2.x.
        at build_dl88ph5x3js4p0j7p18rbzvjq.run(/.../java-client/build.gradle:34)".

* Use pre-configured mavenCentral() repository

https://docs.gradle.org/4.4.1/release-notes.html#use-of-canonical-url-for-mavencentral()-repository-url