Calling NPM without the user's shell profile.
According to the NPM man page, npm gets its configuration information from 5 separate sources. When npm is launched on a *nix system without the user shell npm skips any configuration in the user's $HOME as well as ignores environment variables that might be set in the user's profile or environment.

This resolves #1889 and @ZuhayrMerchant110's issue as well as mine where the project directory prefixing the appium path.