Update IOSDriver.java
Removed the html5 package as this package is no longer supported and deprecated in Selenium 4. Due to this we face compilation issue for java-client wherever it is used. For now I am submitting the PR for IOSDriver. Please check for other related files where html5 is being used and remove the imports.